requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
typing-extensions
python-multipart
python-dotenv
pymysql